moveout LinkedColorList
